django_testing/requirements.txt
asgiref==3.7.2
django-filter==23.3
Django==4.2.5
djangorestframework==3.14.0
exceptiongroup==1.1.3
iniconfig==2.0.0
model-bakery==1.15.0
packaging==23.1
pip==21.3.1
pluggy==1.3.0
psycopg2-binary==2.9.7
pytest-django==4.5.2
pytest==7.4.2
pytz==2023.3.post1
setuptools==60.2.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.8.0
wheel==0.37.1
